Support more than INT_MAX non-zero elements in sparse matrices.
This change allows ceres to solve larger sparse problems.

It is an intermediate step towards 64bit index support for sparse matrices (which would allow ceres to solve even larger problems and also allow for GPU acceleration in SuiteSparse).

This change does not switch to 64bit cell indices though, so the matrix size is still limited to INT_MAX elements per dimension.

However, values used to index into value and row/col arrays are promoted to 64-bit signed integers, so sparse matrices (except for CompressedRowSparseMatrix) can now have more than INT_MAX nonzero elements.

+++ internal/ceres/sparse_matrix.h
@@ -100,7 +100,7 @@ class CERES_EXPORT_INTERNAL SparseMatrix : public LinearOperator {
 
   virtual int num_rows() const = 0;
   virtual int num_cols() const = 0;
-  virtual int num_nonzeros() const = 0;
+  virtual int64_t num_nonzeros() const = 0;
 };
 
 }  // namespace internal
